+++ src/mod.ts
@@ -8,7 +8,7 @@
 import { walkSync } from "@std/fs";
 import { relative } from "@std/path";
 import { kuusiConfig } from "./config.ts";
-import { type Route, WebHook, WebSource } from "./types.ts";
+import { Route, WebHook, WebSource } from "./types.ts";
 import {
   getAmbiguousURLs,
   getDuplicate,
@@ -23,19 +23,10 @@ import {
 export * from "./env.ts";
 export * from "./types.ts";
 
-function checkDuplicateRoutes(routes: Route[]): void {
-  const parsedURLs = routes.map(([url]) =>
-    url.pathname.replace(/\/:[^\/]+(?!=\/)/g, "®")
+const noRouteExport = (path: string): Error =>
+  new Error(
+    `kuusi-invalid-route-export: The route file ${path} does not provide a valid default route export.`,
   );
-
-  const duplicate = getDuplicate(parsedURLs)[0];
-
-  if (duplicate) {
-    throw new Error(
-      `kuusi-duplicate-routes: The "${duplicate}" URL is served multiple times.`,
-    );
-  }
-}
 
 /**
  * Function that collects all the routes from the routes directory.
@@ -43,46 +34,58 @@ function checkDuplicateRoutes(routes: Route[]): void {
  * @returns {Promise<Route[]>} The routes and their paths as KuusiRoutes
  */
 export async function getKuusiRoutes(): Promise<Route[]> {
-  const paths = Array.from(
-    walkSync(kuusiConfig.routes.path, { includeDirs: false }),
-    ({ path }) => relative(kuusiConfig.routes.path, path),
-  );
-
   const routes: Route[] = [];
+  const directoryPath = kuusiConfig.routes.directoryPath;
 
-  for (const path of paths) {
-    if (!routeGuard(path)) continue;
+  if (directoryPath) {
+    const paths = Array.from(
+      walkSync(directoryPath, { includeDirs: false }),
+      ({ path }) => relative(directoryPath, path),
+    );
 
-    const absolutePath = toLocalPath(kuusiConfig.routes.path, path).href;
-    const imports = await import(absolutePath) as object;
+    for (const path of paths) {
+      if (!routeGuard(path)) continue;
 
-    if (!("default" in imports)) {
-      throw new Error(
-        `kuusi-no-route-export: ${absolutePath} does not provide a default export`,
+      const absolutePath = toLocalPath(directoryPath, path).href;
+      const imports = await import(absolutePath);
+
+      if (
+        (sourceGuard(path) && !(imports.default instanceof WebSource)) ||
+        (hookGuard(path) && !(imports.default instanceof WebHook))
+      ) throw noRouteExport(path);
+
+      routes.push(
+        new Route({
+          urlPattern: new URLPattern({ pathname: parsePath(path) }),
+          route: imports.default as WebSource | WebHook,
+        }),
       );
     }
-
-    if (sourceGuard(path)) {
-      if (!(imports.default instanceof WebSource)) {
-        throw new Error(
-          `kuusi-no-source-export: ${absolutePath} does not provide a source export`,
-        );
-      }
-    } else if (hookGuard(path)) {
-      if (!(imports.default instanceof WebHook)) {
-        throw new Error(
-          `kuusi-no-hook-export: ${absolutePath} does not provide a webhook export`,
-        );
-      }
-    } else continue;
-
-    routes.push([
-      new URLPattern({ pathname: parsePath(path) }),
-      imports.default as WebSource | WebHook,
-    ]);
   }
 
-  checkDuplicateRoutes(routes);
+  for (const path of kuusiConfig.routes.filePaths) {
+    const routeFileImports = await import(toLocalPath(path).href) as object;
+
+    if (Object.keys(routeFileImports).length === 0) throw noRouteExport(path);
+
+    for (const routeFileImport of Object.values(routeFileImports)) {
+      if (routeFileImport instanceof Route) routes.push(routeFileImport);
+    }
+  }
+
+  /* checkDuplicateRoutes */ {
+    const parsedURLs = routes.map(({ urlPattern }) =>
+      urlPattern.pathname.replace(/\/:[^\/]+(?!=\/)/g, "®")
+    );
+
+    const duplicate = getDuplicate(parsedURLs)[0];
+
+    if (duplicate) {
+      throw new Error(
+        `kuusi-duplicate-routes: The ${duplicate} URL is served multiple times.`,
+      );
+    }
+  }
 
   if (kuusiConfig.routes.warnAmbiguousRoutes) {
     for (const ambiguousURL of getAmbiguousURLs(routes)) {
@@ -92,7 +95,17 @@ export async function getKuusiRoutes(): Promise<Route[]> {
     }
   }
 
-  return routes;
+  // Sorting and reversing makes sure the pathnames are checked in the
+  // following order:
+  // 1. Normal pathnames
+  // 2. Generic pathnames (those starting with a colon)
+  // 3. The root pathname
+  return routes.sort((a, b) => {
+    const pathA = a.urlPattern.pathname.toLowerCase();
+    const pathB = b.urlPattern.pathname.toLowerCase();
+
+    return pathA.localeCompare(pathB);
+  }).reverse();
 }
 
 /**
@@ -110,7 +123,7 @@ export async function getKuusiRoutes(): Promise<Route[]> {
 export async function kuusi(req: Request, routes: Route[]): Promise<Response> {
   const headers = {
     "content-type": "application/json; charset=utf-8",
-  };
+  } as const;
 
   if (!httpVerbs.includes(req.method)) {
     return new Response("{}", {
@@ -119,7 +132,7 @@ export async function kuusi(req: Request, routes: Route[]): Promise<Response> {
     });
   }
 
-  const match = routes.find(([url]) => url.test(req.url));
+  const match = routes.find(({ urlPattern }) => urlPattern.test(req.url));
 
   if (!match) {
     return new Response("{}", {
@@ -128,14 +141,12 @@ export async function kuusi(req: Request, routes: Route[]): Promise<Response> {
     });
   }
 
-  const [matchPattern, matchRoute] = match;
-  const matchPatternResult = matchPattern.exec(req.url)!;
-  const matchMethod = matchRoute[req.method as keyof (WebSource | WebHook)];
+  const { urlPattern, route } = match;
+  const patternResult = urlPattern.exec(req.url)!;
+  const method = route[req.method as keyof (WebSource | WebHook)];
 
-  return matchMethod
-    ? await matchMethod(req, matchPatternResult)
-    : new Response("{}", {
-      status: 405,
-      ...headers,
-    });
+  return method ? await method(req, patternResult) : new Response("{}", {
+    status: 405,
+    ...headers,
+  });
 }