+++ src/utils.ts
@@ -18,29 +18,17 @@ export const httpVerbs: string[] = [
   "OPTIONS",
 ] as const;
 
-export const ObjectEntriesof = <T extends object>(obj: T) =>
-  Object.entries(obj) as [keyof T, T[keyof T]][];
-
-export const isObjKey = <T extends NonNullable<object>>(
-  key: string | number | symbol,
-  obj: T,
-): key is keyof T => key in obj;
-
-export const isObjField = <T extends NonNullable<object>>(
-  key: string | number | symbol,
-  value: unknown,
-  obj: T,
-): key is keyof T => isObjKey(key, obj) && typeof value === typeof obj[key];
-
 export const parsePath = (path: string) => {
   // Removes the file extensions
   path = path.split(".").slice(0, -1).join(".");
 
+  // Removes the .hook or .source
   if (path.endsWith("hook")) path = path.slice(0, -5);
   else if (path.endsWith("source")) path = path.slice(0, -7);
 
   if (path.endsWith("index")) path = path.slice(0, -5);
 
+  // : is reserved, so use ; instead
   if (Deno.build.os === "windows") path.replace(/\;/g, ";");
 
   return "/" + path;
@@ -51,8 +39,10 @@ export const getDuplicate = <T>(array: T[]) =>
 
 export const getAmbiguousURLs = (routes: Route[]) =>
   getDuplicate(
-    routes.map(([url]) =>
-      url.pathname.endsWith("/") ? url.pathname.slice(0, -1) : url.pathname
+    routes.map(({ urlPattern }) =>
+      urlPattern.pathname.endsWith("/")
+        ? urlPattern.pathname.slice(0, -1)
+        : urlPattern.pathname
     ),
   );