+++ tests/src.test.ts
@@ -1,6 +1,7 @@
 import { assert, assertEquals } from "@std/assert";
 import { assertFalse } from "@std/assert/false";
 import { parsePath, routeGuard } from "../src/utils.ts";
+import { KuusiConfig } from "@kuusi/kuusi/types";
 
 Deno.test({
   name: "Parse Path",
@@ -11,6 +12,7 @@ Deno.test({
     assertEquals(parsePath("index.hook.ts"), "/");
     assertEquals(parsePath("kuusi/index.hook.cts"), "/kuusi/");
     assertEquals(parsePath("kuusi.hook.mts"), "/kuusi");
+    assertEquals(parsePath("viisi/:id.source.ctsx"), "/viisi/:id");
   },
 });
 
@@ -28,5 +30,37 @@ Deno.test({
     assertFalse(routeGuard("yksi.hook.kts"));
     assertFalse(routeGuard("kaksi.hook.sqwuimble"));
     assertFalse(routeGuard("kolme.source.kmtsx"));
+  },
+});
+
+Deno.test({
+  name: "Kuusi Config",
+  fn: () => {
+    const config = new KuusiConfig({
+      routes: {
+        directoryPath: "routes",
+        warnAmbiguousRoutes: true,
+        filePaths: [],
+      },
+      dotenv: {
+        path: ".env",
+        requiredPath: "required.env",
+        export: false,
+        requiredKeys: [],
+      },
+    });
+    assertEquals(new KuusiConfig(), config);
+    assertEquals(new KuusiConfig({}), config);
+    assertEquals(new KuusiConfig({ routes: {} }), config);
+    config.dotenv.export = true;
+    assertEquals(new KuusiConfig({ dotenv: { export: true } }), config);
+    config.routes.filePaths = ["haiii", ":3"];
+    assertEquals(
+      new KuusiConfig({
+        dotenv: { export: true },
+        routes: { filePaths: ["haiii", ":3"] },
+      }),
+      config,
+    );
   },
 });